Changed the code to consider negative strand data.
Implemented a function that indexes the genome (in py dictionaries) by
chromosome and then by gene id.

Now density is calculated independently for the two strands and finally
added up.

Added the skeleton for a better random sampler.

Ran a couple of experiments to see the metagene plots for positive
strand only, negative strand only and both combined.

Script segfaults on chrVII. Maybe it's because RAM limitation on a
laptop? Will run on jr.